S3 read functions {foreign}	R Documentation

Read an S3 Binary or data.dump File

Description

Reads binary data files or data.dump files that were produced in S version 3.

Usage


  data.restore(file, print = FALSE, verbose = FALSE, env = .GlobalEnv)
  read.S(file)

Arguments

file
the filename of the S-PLUS data.dump or binary file.

print
whether to print the name of each object as read from the file.

verbose
whether to print the name of every subitem within each object.

env
environment within which to create the restored object(s).

Details

read.S can read the binary files produced in some older versions of S-PLUS on either Windows (versions 3.x, 4.x, 2000) or Unix (version 3.x with 4 byte integers). It automatically detects whether the file was produced on a big- or little-endian machine and adapts itself accordingly.

data.restore can read a similar range of files produced by data.dump and for newer versions of S-PLUS, those from data.dump(....., oldStyle=TRUE).

Not all S3 objects can be handled in the current version. The most frequently encountered exceptions are functions and expressions; you will also have trouble with objects that contain model formulas. In particular, comments will be lost from function bodies, and the argument lists of functions will often be changed.

Value

For read.S, an R version of the S3 object.

For data.restore, the name of the file.

Author(s)

Duncan Murdoch

Examples

## if you have an S-PLUS _Data file containing 'myobj'
## Not run: read.S(file.path("_Data", "myobj"))
data.restore("dumpdata", print = TRUE)

## End(Not run)

[Package foreign version 0.8-71 Index]