read.mtp {foreign}	R Documentation

Read a Minitab Portable Worksheet

Description

Return a list with the data stored in a file as a Minitab Portable Worksheet.

Usage

read.mtp(file)

Arguments

file
character variable with the name of the file to read. The file must be in Minitab Portable Worksheet format.

Value

A list with one component for each column, matrix, or constant stored in the Minitab worksheet.

Note

This function was written around 1990 for the format current then. Later versions of Minitab appear to have added to the format.

Author(s)

Douglas M. Bates

References

http://www.minitab.com/

Examples

## Not run: 
read.mtp("ex1-10.mtp")

## End(Not run)

[Package foreign version 0.8-65 Index]