lookup.xport {foreign}	R Documentation

Lookup Information on a SAS XPORT Format Library

Description

Scans a file as a SAS XPORT format library and returns a list containing information about the SAS library.

Usage

lookup.xport(file)

Arguments

file 	 character variable with the name of the file to read. The file must be in SAS XPORT format.

Value

A list with one component for each dataset in the XPORT format library.

Author(s)

Saikat DebRoy

References

SAS Technical Support document TS-140: “The Record Layout of a Data Set in SAS Transport (XPORT) Format” available as http://ftp.sas.com/techsup/download/technote/ts140.html.

See Also

read.xport

Examples

## Not run: 
lookup.xport("transport")
## End(Not run)

[Package foreign version 0.8-37 Index]