read.arff {foreign}	R Documentation

Read Data from ARFF Files

Description

Reads data from Weka Attribute-Relation File Format (ARFF) files.

Usage

read.arff(file)

Arguments

file 	 a character string with the name of the ARFF file to read from, or a connection which will be opened if necessary, and if so closed at the end of the function call.

Value

A data frame containing the data from the ARFF file.

References

Attribute-Relation File Format http://www.cs.waikato.ac.nz/~ml/weka/arff.html
http://weka.sourceforge.net/wekadoc/index.php/en:ARFF_(3.5.1)

See Also

write.arff

[Package foreign version 0.8-37 Index]