read.epiinfo {foreign}	R Documentation

Read Epi Info Data Files

Description

Reads data files in the .REC format used by Epi Info versions 6 and earlier and by EpiData. Epi Info is a public domain database and statistics package produced by the US Centers for Disease Control and EpiData is a freely available data entry and validation system.

Usage

read.epiinfo(file, read.deleted = FALSE, guess.broken.dates = FALSE,
             thisyear = NULL, lower.case.names = FALSE)

Arguments

file 	 A filename, URL, or connection.

read.deleted 	 Deleted records are read if TRUE, omitted if FALSE or replaced with NA if NA.

guess.broken.dates 	 Attempt to convert dates with 0 or 2 digit year information (see ‘Details’).

thisyear 	 A 4-digit year to use for dates with no year. Defaults to the current year.

lower.case.names 	 Convert variable names to lowercase?

Details

Epi Info allows dates to be specified with no year or with a 2 or 4 digits. Dates with four-digit years are always converted to Date class. With the guess.broken.dates option the function will attempt to convert two-digit years using the operating system's default method (see Date) and will use the current year or the thisyear argument for dates with no year information.

If read.deleted is TRUE the "deleted" attribute of the data frame indicates the deleted records.

Value

A data frame.

Note

Epi Info 2000, the current version, uses the Microsoft Access file format to store data. This may be readable with the RODBC or RDCOM packages.

References

http://www.cdc.gov/epiinfo/, http://www.epidata.dk

See Also

DateTimeClasses

Examples

## Not run: 
data<-read.epiinfo("oswego.rec",guess.broken.dates=TRUE,thisyear="1972")
## End(Not run)

[Package foreign version 0.8-37 Index]