read.octave {foreign}	R Documentation

Read Octave Text Data Files

Description

Read a file in Octave text data format into a list.

Usage


read.octave(file)

Arguments

file
a character string with the name of the file to read.

Details

This function is used to read in files in Octave text data format, as created by save -text in Octave. It knows about most of the common types of variables, including the standard atomic (real and complex scalars, matrices, and N-d arrays, strings, ranges, and boolean scalars and matrices) and recursive (structs, cells, and lists) ones, but has no guarantee to read all types. If a type is not recognized, a warning indicating the unknown type is issued, it is attempted to skip the unknown entry, and NULL is used as its value. Note that this will give incorrect results, and maybe even errors, in the case of unknown recursive data types.

As Octave can read MATLAB binary files, one can make the contents of such files available to R by using Octave's load and save (as text) facilities as an intermediary step.

Value

A list with one named component for each variable in the file.

Author(s)

Stephen Eglen stephen@gnu.org and Kurt Hornik

References

http://www.octave.org/

[Package foreign version 0.8-71 Index]