write.arff {foreign}	R Documentation

Write Data into ARFF Files

Description

Writes data into Weka Attribute-Relation File Format (ARFF) files.

Usage

write.arff(x, file, eol = "\n")

Arguments

x 	 the data to be written, preferably a matrix or data frame. If not, coercion to a data frame is attempted.

file 	 either a character string naming a file, or a connection. "" indicates output to the standard output connection.

eol 	 the character(s) to print at the end of each line (row).

References

Attribute-Relation File Format http://www.cs.waikato.ac.nz/~ml/weka/arff.html

See Also

read.arff

Examples

write.arff(iris, file = "")

[Package foreign version 0.8-29 Index]